requirements.txt
joblib
pickle-mixin
pandas
numpy
matplotlib
seaborn
plotly
scikit-learn==1.3.2
imbalanced-learn
xgboost
python-dotenv
mysql-connector-python